requirements.txt
asgiref==3.8.1
certifi==2023.5.7
cffi==1.17.1
charset-normalizer==2.0.12
contourpy==1.3.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
cycler==0.12.1
Django==3.2.20
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.8
django-ckeditor==6.2.0
django-js-asset==2.2.0
django-session-timeout==0.1.0
djangorestframework==3.14.0
fonttools==4.55.0
gunicorn==23.0.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
numpy==2.1.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
requests==2.27.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
whitenoise==6.8.2
